+++ onnxruntime/python/tools/kernel_explorer/kernels/gemm_rocblas.h
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include <pybind11/pybind11.h>
+
+namespace py = pybind11;
+
+namespace onnxruntime {
+
+void InitRocBlasGemm(py::module mod);
+
+}
